// FE.Core.FW/src/apis/models/TestScoreModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type TestScoreModel = {
  id?: string;
  firstName?: string | null;
  lastName?: string | null;
  fullName?: string | null;
  dob?: string;
  idOrPassport?: string | null;
  listening?: number;
  reading?: number;
  total?: number;
  testDate?: string;
  formCode?: string | null;
};